contrib/barbican/requirements.txt
python-barbicanclient=2.0.0
